done save transactions and I have a problem get transaction on system
