+++ popup/selectBank.html
@@ -27,6 +27,6 @@
         </div>
       </div>
     </div>
-    <script type="module" src="/js/bank.js"></script>
+    <script type="module" src="/js/selectBank.js"></script>
   </body>
 </html>
